requirements.txt
langchain-openai>=0.0.24
langgraph>=0.0.19
langsmith>=0.1.2
neo4j>=5.17
rdkit-pypi
guacamol==0.5.4
numpy
pandas
pyyaml
tqdm